api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
openai==1.30.1
mediapipe==0.10.14
Pillow==10.3.0
pydantic==2.7.1
pydantic-settings==2.3.0
python-dotenv==1.0.1
